feature: 业务集改造-问题修复、代码重构 #728
